+++ occ-raycasting-cli/src/options.rs
@@ -1,4 +1,4 @@
-use clap::{Parser, Subcommand, ValueEnum};
+use clap::{Parser, ValueEnum};
 use log::{info, LevelFilter};
 
 /// Workaround for parsing the different log level
@@ -31,19 +31,9 @@ pub struct Options {
     #[arg(short, value_enum, long, default_value_t = LogLevel::Info)]
     pub log_level: LogLevel,
 
-    /// The input files
+    /// The test configuration file
     #[arg(short, long)]
-    pub input_files: String,
-
-    /// The occlusion test subcommand
-    #[command(subcommand)]
-    pub occ: OccTestSubcommand,
-}
-
-#[derive(Subcommand, Debug, Clone)]
-pub enum OccTestSubcommand {
-    /// Using a simple rasterizer
-    Rasterizer(RasterizerOptions),
+    pub config: String,
 }
 
 /// The arguments for the rasterizer occlusion test
@@ -58,13 +48,6 @@ impl Options {
     /// Dumps the options to the log.
     pub fn dump_to_log(&self) {
         info!("Log Level: {:?}", self.log_level);
-        info!("Input files: {:?}", self.input_files);
-
-        match &self.occ {
-            OccTestSubcommand::Rasterizer(options) => {
-                info!("Occ Test: Rasterizer");
-                info!("Image Size: {:?}", options.image_size);
-            }
-        }
+        info!("Config file: {:?}", self.config);
     }
 }