+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added loader for CAD files.
 - Added occlusion culling based on rasterization.
+- Occlusion test setups are now provided in a toml configuration file.
 
 ### Changed